Subject: Add project.nvim and its telescope extension to the plugin list
In this commit, the project.nvim, a neovim plugin developed by ahmedkhalf, has been added to the lazy plugin setup. This plugin is dependent on the nvim-telescope/telescope.nvim. The telescope extension named "projects" has also been loaded as part of the setup. The default settings of the project_nvim have been used in the configuration. Any specific configuration can be added in the provided space.