Subject: Remove unused coc_nvim plugin from lua plugin setup
In the Lua plugin setup, the coc_nvim plugin was found to be unused. This plugin used to provide powerful, extensible Vim/Neofor autocompletion, linting, and language server protocol support. However, it is no longer necessary and has been removed from the setup.

This commit removes the call to require the coc_nvim plugin in plugin.lua, and deletes the coc_nvim.lua file which contained the plugin's configuration.