Subject: lua/plugin.lua: Comment out experimental plugins
Body:
The chrisgrieser/nvim-early-retirement and hinell/lsp-timeout.nvim plugins have been commented out in the lua/plugin.lua file. This change is likely due to issues with these plugins or a decision to temporarily disable them. Further adjustments may be necessary based on future requirements or updates to these plugins.