Subject: Simplify comment_nvim setup and remove nvim-ts-context-commentstring dependency
Body:
This commit simplifies the setup of the Comment.nvim plugin in the comment_nvim.lua file. Instead of setting a pre_hook that returns vim.bo.commentstring, we now simply call the setup function with an empty table.

In addition, the dependency on JoosepAlviste/nvim-ts-context-commentstring in the nvim_treesitter.lua file has been removed. This is inline with the simplification of the Comment.nvim setup, as it appears the context commentstring module is no longer needed.

The lazy loading event for comment_nvim has been also changed to false.